% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDS.R
\name{plotMDS}
\alias{plotMDS}
\title{Plot MDS for a DESeq object}
\usage{
plotMDS(rld)
}
\description{
Plot MDS for a DESeq object
}
\examples{
library(DESeq2)
library(ggplot2)
library(cowplot)
dds <- makeExampleDESeqDataSet(betaSD=1)
rld <- rlog(dds)
plotMDS(rld)
}
