% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDS.R
\name{plotMDS}
\alias{plotMDS}
\title{PlotMDS for a DESeq2 object}
\usage{
plotMDS(rld)
}
\arguments{
\item{rld}{A \code{\link{DESeqTransform}} object from DESeq2 library}
}
\description{
This function takes a DESeqTRansform object and performs MDS transformation on it to plot it
}
\examples{
library("DESeq2")
library("ggplot2")
library("cowplot")
dds <- makeExampleDESeqDataSet(betaSD=1)
rld <- rlog(dds)
plotMDS(rld)

}
\author{
Cristóbal Fresno
}
